test: split example and property tests
